program: Implement rent collection.
This will send the remaining balance on the escrow account (rent) to
a hardcoded address, which can be considered as a fee for using the
service successfully.

Note that this does not apply when the stream is cancelled, as then we
shall simply return the rent to the stream initializer.